+++ Changes.md
@@ -1,5 +1,11 @@
 ﻿# Change log
 
+## 1.0.7
+* Added `avgColor` (the average color of the photo) to the `Photo` object output
+* Added `orientation`, `size` and `color` to the Photo Search parameters.
+* Added `orientation`, `size` and `locale` to the Video Search parameters.
+* Removed deprecated `minWidth`, `minHeight`, `minDuration`, `maxDuration` parameters from Video Search.
+
 ## 1.0.6
 * Add support for returning collections belonging to the API user.
 * Add support for returning media from a collection.